- Show the total worth of the inventory in the client's UI - Remove singleton logic layer between inventory service and data sink layer
