+++ GildedRose.Server/Startup.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
@@ -10,6 +9,7 @@ namespace GildedRose.Server
     {
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddSingleton<InventoryService>();
             services.AddGrpc();
         }
